feat: 콘솔 애플리케이션 레포지토리 구현
